Changed addFact
The update_KB_is returns true if the KB is updated and false if the knowledge is already known. It still updates the KB as expected. I changed addFact() as well. I am not exactly sure what messages we want to send during addFacts, so I made it work like it did before as much as possible.